--- client/src/pages/SignIn.module.scss
@@ -1,39 +1,0 @@
-@use '../styles/color' as C;
-@use '../styles/utils' as U;
-@use '../styles/typography' as T;
-
-.container {
-  padding: 96px;
-  @include U.flex-c(column, flex-start, flex-start);
-  gap: 24px;
-}
-
-.label {
-  @include T.black-regular-14;
-  @include U.flex-c(column, flex-start, flex-start);
-  gap: 12px;
-}
-
-.inputErrorBox {
-  @include U.flex-c(row, flex-start, center);
-  gap: 10px;
-
-  input {
-    padding-left: 10px;
-    border: 1px solid C.$gray300;
-    border-radius: 6px;
-    @include U.size(240px, 40px);
-  }
-}
-
-.errorMessage {
-  @include T.blue-regular-12;
-}
-
-.submitBtn {
-  margin-top: 24px;
-  background-color: C.$blue400;
-  border-radius: 6px;
-  color: C.$white;
-  padding: 10px 16px;
-}
